+++ silverstrike/__init__.py
@@ -1,2 +1,2 @@
-__version__ = '0.1.1'
+__version__ = '0.1.3'
 VERSION = __version__